test(inspector): add missing ArtifactRichTextCodec source + delegation (#710)
Follow-up to ed30d631, which registered ArtifactRichTextCodec.swift in the
project and added the tests but omitted the source file itself and the
ArtifactPanel delegation edit (bad staging path). This adds them so the
codec the tests + pbxproj reference actually exists.
